requirements.txt
asyncpg==0.23.0
python-dateutil==2.8.2
fastapi==0.66.0
pydantic==1.8.2
python-dotenv==0.18.0
uvicorn==0.14.0
